my own simple tokenizer -- no problems with missing parentheses or glued punctuators and taking function code directly (no need to restart server)

+++ controllers/lesson0_php_to_py.py
@@ -21,4 +21,4 @@ def functions():
         result = x*2    ###PLACEHOLDER:-->    $result = $x*2;
         return result   ###PLACEHOLDER:-->    return $resul;
                         ###PLACEHOLDER:--> }
-    return double( 5 )
+    return double( 7 )
